refactor : pages 컴포넌트명에 'page' *-postfix 추가
