+++ src/router.tsx
@@ -1,15 +1,15 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { PATH } from "./constants/path";
-import { AddCard } from "./pages/addCard";
-import { CardList } from "./pages/cardList";
-import { ErrorPage } from "./pages/error";
+import { AddCardPage } from "./pages/addCardPage";
+import { CardListPage } from "./pages/cardListPage";
+import { ErrorPage } from "./pages/errorPage";
 
 export function Router() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path={PATH.CARD_LIST} element={<CardList />}></Route>
-        <Route path={PATH.ADD_CARD} element={<AddCard />}></Route>
+        <Route path={PATH.CARD_LIST} element={<CardListPage />}></Route>
+        <Route path={PATH.ADD_CARD} element={<AddCardPage />}></Route>
         <Route path="*" element={<ErrorPage />} />
       </Routes>
     </BrowserRouter>
